removing software safety projections
I have a slightly different controller that expects servo values to be specified between 400 and 2200.  The posiitonal limitation from 0-1000 did not provide ample coverage of this range.  I considered making the range configurable so that floats could always be from -125.0 to 125.0 and there could be some controller level configuration that would specify the range... But... I don't have enough time to figure that out currently.

Additonally, I'm having issues getting the servo states, but, thats a problem for another day.